organizating code in internal private and public sdk section
